chore: rename component folder
