+++ frontend/src/Root.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { Header } from "./components/composed/Header";
+import { Header } from "./components/composed_ui/Header";
 
 export function Root() {
   return (
